requirements.txt
Django>=5.0,<6.0
djangorestframework
speechrecognition
pydub
numpy
pandas
python-dotenv